Room adjustments

+++ app/src/main/java/com/example/kotlinbaseboilerplate/data/db/entity/WeatherLocation.kt
@@ -15,7 +15,7 @@ data class WeatherLocation(
     @SerializedName("localtime")
     val localtime: String,
     @SerializedName("localtime_epoch")
-    val localtimeEpoch: Int,
+    val localtimeEpoch: Long,
     @SerializedName("lon")
     val lon: String,
     @SerializedName("name")
